requirements.txt
mcp>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
streamlit>=1.28.0
deep-translator>=1.11.0
python-dotenv>=1.0.0

